comments + trying to fix jenkins test failures
